解决tk.mapper和之前代码之间的一些冲突

+++ analyse-service/src/main/java/com/bolean/dao/FolderMapper.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 
 public interface FolderMapper extends BaseMapper<Folder> {
-    List<Folder> selectByParentId(Integer parentId);
+    List<Folder> selectByParentId(Long parentId);
 
     List<Folder> selectByRoleFolders(Map<String,Object> map);
 }
