+++ analyse-service/src/main/java/com/bolean/dao/RoleFolderMapper.java
@@ -4,5 +4,5 @@ import com.bolean.entity.RoleFolder;
 import mybatis.basemapper.BaseMapper;
 
 public interface RoleFolderMapper extends BaseMapper<RoleFolder> {
-    int deleteByRoleId(Integer roleId);
+    int deleteByRoleId(Long roleId);
 }
